moved commands to subdirectories
