+++ src/lib/commands/file-operations/index.js
@@ -1,0 +1,4 @@
+export { default as addCommand } from './add-command.js';
+export { default as catCommand } from './cat-command.js';
+export { default as cpCommand } from './cp-command.js';
+export { default as rnCommand } from './rn-command.js';
