+++ src/lib/commands/os-info/index.js
@@ -1,0 +1,1 @@
+export { default as osCommand } from './os-command.js';
