+++ src/lib/commands/working-directory-navigation/index.js
@@ -1,0 +1,3 @@
+export { default as cdCommand } from './cd-command.js';
+export { default as lsCommand } from './ls-command.js';
+export { default as upCommand } from './up-command.js';
