+++ src/lib/commands/working-directory-navigation/up-command.js
@@ -1,5 +1,5 @@
 import path from 'node:path';
-import InvalidInputError from '../custom-errors/invalid-input-error.js';
+import InvalidInputError from '../../custom-errors/invalid-input-error.js';
 
 const upCommand = {
     commandName: 'up',
